Rename `domainSearch` reducer `domainSuggestions`
